+++ config/default/userService.conf.php
@@ -1,0 +1,3 @@
+<?php
+
+return new \oat\taoOauth\model\user\UserService();
